% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Odds Ratio and 95% Confidence Interval Calculation}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{Numeric value representing the coefficient.}

\item{se}{Numeric value representing the standard error.}

\item{siglevel}{Numeric value specifying the significance level for the CI.}

\item{roundto}{Integer specifying the number of decimal places to round the results to.}
}
\value{
A character string representing the OR and its 95% CI in the format "OR (Lower CI, Upper CI)".
}
\description{
Calculates the odds ratio (OR) and its 95% confidence interval (CI)
  based on the given coefficients and standard error.
}
\examples{
OR_95CI(0.5, 0.2, 0.05, 2)
}
\author{
Sarra Ali
}
