% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Description of Data.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data: toydata}
\format{
A data frame with 100 rows and 3 variables.
\describe{
  \item{y}{Binary outcome.}
  \item{x1}{Continuous predictor.}
  \item{x2}{Binary predictor.}
}
}
\usage{
toydata
}
\description{
This is an example dataset for demonstration purposes. It contains three columns:
- `y`: a binary outcome
- `x1`: a continuous predictor
- `x2`: a binary predictor
}
\keyword{datasets}
